refactor(cli): i18n-dict 按 command/domain 拆分为子文件
把 768 行的 src/cli/lib/i18n-dict.ts 按命令/领域拆为 6 个子文件
(init / run / gen / evolve / help / common),i18n-dict.ts 退为
60 行 barrel:类型 union + dict 合并,外部接口
(CLI_DICT / CliMessageKey / CliMessage) 完全不变。

子文件结构: 每个 *.ts 暴露自己的 *MessageKey union + *Dict
Record;新增文案直接编辑对应文件即可,不再争夺单一巨文件。
i18n-dict/types.ts 单独持有 CliMessage interface。

无生产语义改动,纯组织结构调整;无 BREAKING-COMPARABILITY,
测量学红区文件零 touch,1615 用例全绿。
