+++ src/cli/lib/i18n-dict/types.ts
@@ -1,0 +1,4 @@
+export interface CliMessage {
+  zh: string;
+  en: string;
+}
